% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync_facility.R
\name{initRsync}
\alias{initRsync}
\title{Initialize Rsync}
\usage{
initRsync(login, password = NULL, pwfile = NULL, socket = NULL,
  tempdir.loc, verbosity = 1, timeout.con = 5, delay = c(5, 10),
  delay2 = delay * 1.5)
}
\arguments{
\item{login}{ssh login in the form user@host}

\item{password}{ssh password}

\item{pwfile}{file whose first line contains ssh password (overrides \code{password})}

\item{socket}{socket file created by ssh for the communication (bypasses time dealy to login). optional.}

\item{tempdir.loc}{directory for temporary files}

\item{verbosity}{verbosity of status messages}

\item{timeout.con}{duration before an ongoing connection attempt is considered to be timed out}

\item{delay}{waiting times between connection attempts}

\item{delay2}{upper limit of waiting times between connection attemps}
}
\value{
A list (think object) with functions to pull and push files and directories, see \code{details}
}
\description{
Creates a list with functions that facilitate pushing and pulling files and directories from a remote machine.
}
\details{
The returned list contains the following functions:
\tabular{ll}{
\code{upSyncFile(srcFile, destFile)}\tab copy a file from the local to the remote machine\cr
\code{downSyncFile(srcFile, destFile)}\tab copy a file from the remote to the local machine\cr
\code{upMoveFile(srcFile, destFile)}\tab move a file from the local to the remote machine\cr
\code{downSyncFile(srcFile, destFile)}\tab move a file from the remote machine to the local machine\cr
\code{upSyncDir(srcDir, destDir)}\tab copy a directory recursively from the local to the remote machine\cr
\code{downSyncDir(srcDir, destDir)}\tab copy a directory recursively from the remote to the local machine\cr
\code{setSocketFile(newPath)}\tab change the path to the socket file (created by ssh)\cr
\code{getSocketFile()}\tab return the path to the socket file\cr
\code{closeCon()}\tab perform clean up operations when rsync is no longer needed
}
}
